fix(dashboard): scope homepage skeleton to a route group so other routes keep the generic fallback
